+++ PlantingMind/PlantingMind/Localization/Localizable.xcstrings
@@ -1,15 +1,6 @@
 {
   "sourceLanguage" : "en",
   "strings" : {
-    "%@" : {
-
-    },
-    "%lld" : {
-
-    },
-    "%lld / 100" : {
-
-    },
     "cancel" : {
       "extractionState" : "manual",
       "localizations" : {
@@ -23,6 +14,22 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "취소"
+          }
+        }
+      }
+    },
+    "change_month" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Change Month"
+          }
+        },
+        "ko" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "기간 변경"
           }
         }
       }
@@ -112,8 +119,22 @@
         }
       }
     },
-    "select date" : {
-
+    "text_limit" : {
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "%lld / 100"
+          }
+        },
+        "ko" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "%lld / 100"
+          }
+        }
+      }
     }
   },
   "version" : "1.0"
