+++ PlantingMind/PlantingMind/StringExtension.swift
@@ -11,4 +11,8 @@ extension String {
     var localized: String {
         NSLocalizedString(self, comment: "")
     }
+    
+    func localized(with arguments: [CVarArg] = []) -> String {
+        return String(format: self.localized, arguments: arguments)
+    }
 }
